Refactored to `custom_widgets.py`
