+++ datashuttle/tui/custom_widgets.py
@@ -1,0 +1,77 @@
+from textual.app import ComposeResult
+from textual.containers import Container
+from textual.reactive import reactive
+from textual.screen import ModalScreen
+from textual.widgets import Button, Checkbox, Static
+
+from datashuttle.configs.canonical_configs import get_datatypes
+
+
+class ErrorScreen(ModalScreen):
+    """
+    A screen for rendering error messages. The border of the
+    central widget is red. The screen does not return any value.
+    """
+
+    def __init__(self, message):
+        super(ErrorScreen, self).__init__()
+
+        self.message = message
+
+    def compose(self) -> ComposeResult:
+        yield Container(
+            Container(
+                Static(self.message, id="errorscreen_message_label"),
+                id="errorscreen_message_container",
+            ),
+            Container(Button("OK"), id="errorscreen_ok_button"),
+            id="errorscreen_top_container",
+        )
+
+    def on_button_pressed(self) -> None:
+        self.dismiss()
+
+
+class DatatypeCheckboxes(Static):
+    """
+    Dynamically-populated checkbox widget for convenient datatype
+    selection during folder creation.
+
+    Attributes
+    ----------
+
+    type_out:
+        List of datatypes selected by the user to be passed to `make_folders`
+        (e.g. "behav" that will be passed to `make-folders`.)
+
+    type_config:
+        List of datatypes supported by NeuroBlueprint
+    """
+
+    type_out = reactive("all")
+
+    def __init__(self):
+        super(DatatypeCheckboxes, self).__init__()
+
+        self.type_config = get_datatypes()
+
+    def compose(self):
+        for type in self.type_config:
+            yield Checkbox(
+                type.title(), id=f"tabscreen_{type}_checkbox", value=True
+            )
+
+    def on_checkbox_changed(self):
+        """
+        When a checkbox is clicked, update the `type_out` attribute
+        with the datatypes to pass to `make_folders` datatype argument.
+        """
+        type_dict = {
+            type: self.query_one(f"#tabscreen_{type}_checkbox").value
+            for type in self.type_config
+        }
+        self.type_out = [
+            datatype
+            for datatype, is_on in zip(type_dict.keys(), type_dict.values())
+            if is_on
+        ]
