+++ datashuttle/tui/tui_app.py
@@ -4,97 +4,23 @@ from time import monotonic
 from textual.app import App, ComposeResult
 from textual.binding import Binding
 from textual.containers import Container
-from textual.reactive import reactive
-from textual.screen import ModalScreen, Screen
+from textual.screen import Screen
 from textual.widgets import (
     Button,
-    Checkbox,
     DirectoryTree,
     Header,
     Input,
     Label,
-    Static,
     TabbedContent,
     TabPane,
 )
 
 from datashuttle import DataShuttle
-from datashuttle.configs.canonical_configs import get_datatypes
-from datashuttle.tui import project_config
+from datashuttle.tui import custom_widgets, project_config
 from datashuttle.utils.folders import get_existing_project_paths_and_names
 
 # RENAME ALL WIDGETS
 # TCSS
-
-
-class ErrorScreen(ModalScreen):
-    """
-    A screen for rendering error messages. The border of the
-    central widget is red. The screen does not return any value.
-    """
-
-    def __init__(self, message):
-        super(ErrorScreen, self).__init__()
-
-        self.message = message
-
-    def compose(self) -> ComposeResult:
-        yield Container(
-            Container(
-                Static(self.message, id="errorscreen_message_label"),
-                id="errorscreen_message_container",
-            ),
-            Container(Button("OK"), id="errorscreen_ok_button"),
-            id="errorscreen_top_container",
-        )
-
-    def on_button_pressed(self) -> None:
-        self.dismiss()
-
-
-class DatatypeCheckboxes(Static):
-    """
-    Dynamically-populated checkbox widget for convenient datatype
-    selection during folder creation.
-
-    Attributes
-    ----------
-
-    type_out:
-        List of datatypes selected by the user to be passed to `make_folders`
-        (e.g. "behav" that will be passed to `make-folders`.)
-
-    type_config:
-        List of datatypes supported by NeuroBlueprint
-    """
-
-    type_out = reactive("all")
-
-    def __init__(self):
-        super(DatatypeCheckboxes, self).__init__()
-
-        self.type_config = get_datatypes()
-
-    def compose(self):
-        for type in self.type_config:
-            yield Checkbox(
-                type.title(), id=f"tabscreen_{type}_checkbox", value=True
-            )
-
-    def on_checkbox_changed(self):
-        """
-        When a checkbox is clicked, update the `type_out` attribute
-        with the datatypes to pass to `make_folders` datatype argument.
-        """
-        type_dict = {
-            type: self.query_one(f"#tabscreen_{type}_checkbox").value
-            for type in self.type_config
-        }
-        self.type_out = [
-            datatype
-            for datatype, is_on in zip(type_dict.keys(), type_dict.values())
-            if is_on
-        ]
 
 
 class TabScreen(Screen):
@@ -152,7 +78,7 @@ class TabScreen(Screen):
                     id="tabscreen_session_input", placeholder="e.g. ses-001"
                 )
                 yield Label("Datatype(s)", id="tabscreen_datatype_label")
-                yield DatatypeCheckboxes()
+                yield custom_widgets.DatatypeCheckboxes()
                 yield Button("Make Folders", id="tabscreen_make_folder_button")
 
             with TabPane("Transfer", id="tabscreen_transfer_tab"):
@@ -318,7 +244,7 @@ class TuiApp(App):
             self.push_screen(TabScreen(self, project))
 
     def show_modal_error_dialog(self, message):
-        self.push_screen(ErrorScreen(message))
+        self.push_screen(custom_widgets.ErrorScreen(message))
 
 
 if __name__ == "__main__":
